@@ -30,7 +30,7 @@
     "@babel/core": "^7.21.8",
     "babel-loader": "^8.3.0",
     "ts-loader": "^6.2.2",
-    "webpack": "^4.46.0",
+    "webpack": "5.89.0",
     "webpack-cli": "^3.3.12",
     "webpack-dev-server": "^3.11.3",
     "webpack-merge": "^4.2.2"
