@@ -1,11 +1,11 @@
-/*!
+/*
  * Copyright The OpenTelemetry Authors
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *      http://www.apache.org/licenses/LICENSE-2.0
+ *      https://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
@@ -13,21 +13,6 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-module.exports = {
-  listenAddress: 'localhost',
-  hostname: 'localhost',
-  browsers: ['ChromeHeadless'],
-  frameworks: ['mocha'],
-  coverageReporter: {
-    type : 'json',
-    subdir: '.',
-    dir : '.nyc_output/'
-  },
-  reporters: ['spec', 'coverage'],
-  files: ['test/index-webpack.ts'],
-  preprocessors: {
-    'test/index-webpack*.ts': ['webpack']
-  },
-  webpackMiddleware: { noInfo: true }
-};
+declare module 'systeminformation/lib/network' {
+  export { networkStats } from 'systeminformation';
+}
