@@ -14,12 +14,15 @@
  * limitations under the License.
  */
 
-import * as SI from 'systeminformation';
-import { NetworkData } from '../types';
+// Import from the network file directly as bundlers trigger the 'osx-temperature-sensor' import in the systeminformation/lib/cpu.js,
+// resulting in the following warning: "Can't resolve 'osx-temperature-sensor'"
+// See https://github.com/open-telemetry/opentelemetry-js-contrib/pull/2071
+import { networkStats } from 'systeminformation/lib/network';
+import type { Systeminformation } from 'systeminformation';
 
 export function getNetworkData() {
-  return new Promise<NetworkData[]>(resolve => {
-    SI.networkStats()
+  return new Promise<Systeminformation.NetworkStatsData[]>(resolve => {
+    networkStats()
       .then(resolve)
       .catch(() => {
         resolve([]);
