@@ -58,7 +58,7 @@ export type ExpressRouter = {
 
 // https://github.com/expressjs/express/blob/main/lib/router/layer.js#L33
 export type ExpressLayer = {
-  handle: Function;
+  handle: Function & Record<string, any>;
   [kLayerPatched]?: boolean;
   name: string;
   params: { [key: string]: string };
